Major refactor for Styles and Components
