+++ .gitignore
@@ -17,3 +17,4 @@ public/
 .vscode/
 *.cache.v3
 *.chat*.md
+REFACTORING_SUMMARY.md
