+++ src/components/blog/index.js
@@ -1,0 +1,3 @@
+// Blog components
+export { default as BlogPostListing } from "./BlogPostListing";
+export { default as BlogPost } from "./BlogPost";
