+++ src/components/common/index.js
@@ -1,0 +1,3 @@
+// Common components
+export { default as Icon } from "./Icon";
+export { default as Seo } from "./Seo";
