+++ src/components/layout/index.js
@@ -1,0 +1,4 @@
+// Layout components
+export { default as Header } from "./Header";
+export { default as CompactHeader } from "./CompactHeader";
+export { default as Footer } from "./Footer";
