+++ src/components/ui/index.js
@@ -1,0 +1,2 @@
+// UI components
+export { default as Showcase } from "./Showcase";
