+++ src/styles/base/_global.scss
@@ -1,0 +1,36 @@
+// =======================================================
+// Global Base Styles
+
+@import "../variables";
+
+body {
+  background-color: $bg-body-color;
+  animation: fade-in 0.6s;
+  color: $color-front;
+}
+
+// Paragraph and Content Block Styles
+p,
+.content-block {
+  margin-bottom: 1rem;
+
+  &.--reduced-margin-bottom {
+    margin-bottom: 1rem;
+  }
+}
+
+// Section spacing
+.section-top-margin {
+  margin-top: 10rem;
+
+  &.--tight {
+    margin-top: 4rem;
+  }
+}
+
+// Resume content
+.content-resume {
+  text-align: center;
+  font-size: 2.7rem;
+  margin-top: 3.5rem;
+}
