+++ src/styles/base/_index.scss
@@ -1,0 +1,3 @@
+// Base styles
+@import "global";
+@import "typography";
