+++ src/styles/components/_icons.scss
@@ -1,0 +1,14 @@
+// =======================================================
+// Icon Styles
+
+.icon {
+  display: inline-block;
+
+  svg {
+    margin-right: 2rem;
+    font-size: inherit;
+    height: 1.3em;
+    overflow: visible;
+    vertical-align: -0.4rem;
+  }
+}
