+++ src/styles/components/_index.scss
@@ -1,0 +1,5 @@
+// Component styles
+@import "header";
+@import "footer";
+@import "post-list";
+@import "icons";
