+++ src/styles/layout/_containers.scss
@@ -1,0 +1,62 @@
+// =======================================================
+// Container Styles
+
+@import "../variables";
+@import "../mixins";
+
+.body-container {
+  .content-container {
+    margin: 0 auto;
+    padding: 0 10px;
+    max-width: 800px;
+
+    @media (min-width: 576px) {
+      max-width: 540px;
+    }
+    @media (min-width: 768px) {
+      max-width: 680px;
+    }
+    @media (min-width: 992px) {
+      max-width: 760px;
+    }
+    @media (min-width: 1200px) {
+      max-width: 800px;
+    }
+
+    a {
+      margin-left: 5px;
+      &:not(.icon) {
+        @include underline(#ffffff, #f99);
+      }
+    }
+
+    blockquote {
+      padding: 0 15px;
+      margin: 0;
+      color: #819198;
+      border-left: 0.3rem solid #dce6f0;
+    }
+  }
+}
+
+.post-list-container {
+  max-width: 100%;
+  margin: 0 auto;
+  padding: 0 10px;
+
+  @media (min-width: 576px) {
+    max-width: 540px;
+  }
+  @media (min-width: 768px) {
+    max-width: 720px;
+  }
+  @media (min-width: 992px) {
+    max-width: 960px;
+  }
+  @media (min-width: 1200px) {
+    max-width: 1140px;
+  }
+  @media (min-width: 1400px) {
+    max-width: 1320px;
+  }
+}
