+++ src/styles/layout/_index.scss
@@ -1,0 +1,2 @@
+// Layout styles
+@import "containers";
