+++ src/styles/utilities/_helpers.scss
@@ -1,0 +1,8 @@
+// =======================================================
+// Utility Helper Classes
+
+// Add any utility classes here that don't fit into components
+// For example: spacing utilities, display utilities, etc.
+
+// Currently all utilities are integrated into base and component files
+// This file is ready for future utility classes
