+++ src/styles/utilities/_index.scss
@@ -1,0 +1,2 @@
+// Utility styles
+@import "helpers";
